Removing timeout before cleanup
This test was creating the newfolder/added.js file after calling test.done().  The first file write resulted in two add events - one for the newfolder and one for the newfolder/added.js.  Without the timeout, the test was finishing before the second write.  Now events are only fired for matching paths, so the test behaves as expected.